+++ packages/www/src/documentation/components/content/badge.mdx
@@ -3,16 +3,16 @@ title: Badge
 propsOf: Badge
 ---
 
-`Badge` can be styled for size, color and it has flags for round border and transparency.
+`Badge` has its default values as size small and background color as purple. The styles can be updated for size and color, and it has flags for `rounded` border and `transparent` background.
 
 ## Badge Sizes
 
 ```jsx
-<Badge m="small">small</Badge>
-<Badge m="small" size="medium" >
+<Badge mr="small">small</Badge>
+<Badge mr="small" size="medium" >
   medium
 </Badge>
-<Badge m="small" size="large" >
+<Badge mr="small" size="large" >
   large
 </Badge>
 ```
@@ -20,33 +20,33 @@ propsOf: Badge
 ## Badge Intent
 
 ```jsx
-<Badge intent="plain" m="small" size="medium" >
+<Badge intent="plain" mr="small" size="medium" >
   plain
 </Badge>
-<Badge intent="positive" m="small" size="medium" >
+<Badge intent="positive" mr="small" size="medium" >
   positive
 </Badge>
-<Badge intent="info" m="small" size="medium" >
+<Badge intent="info" mr="small" size="medium" >
   info
 </Badge>
-<Badge intent="neutral" m="small" size="medium" >
+<Badge intent="neutral" mr="small" size="medium" >
   neutral
 </Badge>
-<Badge intent="warning" m="small" size="medium" >
+<Badge intent="warning" mr="small" size="medium" >
   warning
 </Badge>
-<Badge intent="critical" m="small" size="medium" >
+<Badge intent="critical" mr="small" size="medium" >
   critical
 </Badge>
 ```
 
-## Badge Flags
+## Badge Extra Flags
 
 ```jsx
-<Badge m="small" size="medium" transparent >
+<Badge mr="small" size="medium" transparent >
   transparent
 </Badge>
-<Badge m="small" round size="medium" >
-round
+<Badge mr="small" rounded size="medium" >
+  rounded
 </Badge>
 ```
